Extract selectors to module for mixin
